feat(src/): Add predctions obj to state
